requirements.txt
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
jupyter==1.0.0
langchain==0.2.16
sentence-transformers==3.0.1
faiss-cpu==1.8.0
hf_xet==0.2.0